Struct aws_sdk_amplify::client::fluent_builders::UpdateBranch

pub struct UpdateBranch<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> UpdateBranch<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<UpdateBranchOutput, SdkError<UpdateBranchError>> where
R::Policy: SmithyRetryPolicy<UpdateBranchInputOperationOutputAlias, UpdateBranchOutput, UpdateBranchError, UpdateBranchInputOperationRetryAlias>,

pub fn app_id(self, inp: impl Into<String>) -> Self

The unique ID for an Amplify app.

pub fn set_app_id(self, input: Option<String>) -> Self

pub fn branch_name(self, inp: impl Into<String>) -> Self

The name for the branch.

pub fn set_branch_name(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

The description for the branch.

pub fn set_description(self, input: Option<String>) -> Self

pub fn framework(self, inp: impl Into<String>) -> Self

The framework for the branch.

pub fn set_framework(self, input: Option<String>) -> Self

pub fn stage(self, inp: Stage) -> Self

Describes the current stage for the branch.

pub fn set_stage(self, input: Option<Stage>) -> Self

pub fn enable_notification(self, inp: bool) -> Self

Enables notifications for the branch.

pub fn set_enable_notification(self, input: Option<bool>) -> Self

pub fn enable_auto_build(self, inp: bool) -> Self

Enables auto building for the branch.

pub fn set_enable_auto_build(self, input: Option<bool>) -> Self

pub fn environment_variables(
self,
k: impl Into<String>,
v: impl Into<String>
) -> Self

Adds a key-value pair to environmentVariables.

To override the contents of this collection use set_environment_variables.

The environment variables for the branch.

pub fn set_environment_variables(
self,
input: Option<HashMap<String, String>>
) -> Self

pub fn basic_auth_credentials(self, inp: impl Into<String>) -> Self

The basic authorization credentials for the branch.

pub fn set_basic_auth_credentials(self, input: Option<String>) -> Self

pub fn enable_basic_auth(self, inp: bool) -> Self

Enables basic authorization for the branch.

pub fn set_enable_basic_auth(self, input: Option<bool>) -> Self

pub fn enable_performance_mode(self, inp: bool) -> Self

Enables performance mode for the branch.

Performance mode optimizes for faster hosting performance by keeping content cached at the edge for a longer interval. When performance mode is enabled, hosting configuration or code changes can take up to 10 minutes to roll out.

pub fn set_enable_performance_mode(self, input: Option<bool>) -> Self

pub fn build_spec(self, inp: impl Into<String>) -> Self

The build specification (build spec) for the branch.

pub fn set_build_spec(self, input: Option<String>) -> Self

pub fn ttl(self, inp: impl Into<String>) -> Self

The content Time to Live (TTL) for the website in seconds.

pub fn set_ttl(self, input: Option<String>) -> Self

pub fn display_name(self, inp: impl Into<String>) -> Self

The display name for a branch. This is used as the default domain prefix.

pub fn set_display_name(self, input: Option<String>) -> Self

pub fn enable_pull_request_preview(self, inp: bool) -> Self

Enables pull request previews for this branch.

pub fn set_enable_pull_request_preview(self, input: Option<bool>) -> Self

pub fn pull_request_environment_name(self, inp: impl Into<String>) -> Self

The Amplify environment name for the pull request.

pub fn set_pull_request_environment_name(self, input: Option<String>) -> Self

pub fn backend_environment_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) for a backend environment that is part of an Amplify app.

pub fn set_backend_environment_arn(self, input: Option<String>) -> Self