Module aws_sdk_amplify::error

Modules

bad_request_exception

See BadRequestException

dependent_service_failure_exception

See DependentServiceFailureException

internal_failure_exception

See InternalFailureException

limit_exceeded_exception

See LimitExceededException

not_found_exception

See NotFoundException

resource_not_found_exception

See ResourceNotFoundException

unauthorized_exception

See UnauthorizedException

Structs

BadRequestException

A request contains unexpected data.

CreateAppError

CreateBackendEnvironmentError

CreateBranchError

CreateDeploymentError

CreateDomainAssociationError

CreateWebhookError

DeleteAppError

DeleteBackendEnvironmentError

DeleteBranchError

DeleteDomainAssociationError

DeleteJobError

DeleteWebhookError

DependentServiceFailureException

An operation failed because a dependent service threw an exception.

GenerateAccessLogsError

GetAppError

GetArtifactUrlError

GetBackendEnvironmentError

GetBranchError

GetDomainAssociationError

GetJobError

GetWebhookError

InternalFailureException

The service failed to perform an operation due to an internal issue.

LimitExceededException

A resource could not be created because service quotas were exceeded.

ListAppsError

ListArtifactsError

ListBackendEnvironmentsError

ListBranchesError

ListDomainAssociationsError

ListJobsError

ListTagsForResourceError

ListWebhooksError

NotFoundException

An entity was not found during an operation.

ResourceNotFoundException

An operation failed due to a non-existent resource.

StartDeploymentError

StartJobError

StopJobError

TagResourceError

UnauthorizedException

An operation failed due to a lack of access.

UntagResourceError

UpdateAppError

UpdateBranchError

UpdateDomainAssociationError

UpdateWebhookError

Enums

CreateAppErrorKind

CreateBackendEnvironmentErrorKind

CreateBranchErrorKind

CreateDeploymentErrorKind

CreateDomainAssociationErrorKind

CreateWebhookErrorKind

DeleteAppErrorKind

DeleteBackendEnvironmentErrorKind

DeleteBranchErrorKind

DeleteDomainAssociationErrorKind

DeleteJobErrorKind

DeleteWebhookErrorKind

GenerateAccessLogsErrorKind

GetAppErrorKind

GetArtifactUrlErrorKind

GetBackendEnvironmentErrorKind

GetBranchErrorKind

GetDomainAssociationErrorKind

GetJobErrorKind

GetWebhookErrorKind

ListAppsErrorKind

ListArtifactsErrorKind

ListBackendEnvironmentsErrorKind

ListBranchesErrorKind

ListDomainAssociationsErrorKind

ListJobsErrorKind

ListTagsForResourceErrorKind

ListWebhooksErrorKind

StartDeploymentErrorKind

StartJobErrorKind

StopJobErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateAppErrorKind

UpdateBranchErrorKind

UpdateDomainAssociationErrorKind

UpdateWebhookErrorKind