Module aws_sdk_amplify::input

Modules

create_app_input

See CreateAppInput

create_backend_environment_input

See CreateBackendEnvironmentInput

create_branch_input

See CreateBranchInput

create_deployment_input

See CreateDeploymentInput

create_domain_association_input

See CreateDomainAssociationInput

create_webhook_input

See CreateWebhookInput

delete_app_input

See DeleteAppInput

delete_backend_environment_input

See DeleteBackendEnvironmentInput

delete_branch_input

See DeleteBranchInput

delete_domain_association_input

See DeleteDomainAssociationInput

delete_job_input

See DeleteJobInput

delete_webhook_input

See DeleteWebhookInput

generate_access_logs_input

See GenerateAccessLogsInput

get_app_input

See GetAppInput

get_artifact_url_input

See GetArtifactUrlInput

get_backend_environment_input

See GetBackendEnvironmentInput

get_branch_input

See GetBranchInput

get_domain_association_input

See GetDomainAssociationInput

get_job_input

See GetJobInput

get_webhook_input

See GetWebhookInput

list_apps_input

See ListAppsInput

list_artifacts_input

See ListArtifactsInput

list_backend_environments_input

See ListBackendEnvironmentsInput

list_branches_input

See ListBranchesInput

list_domain_associations_input

See ListDomainAssociationsInput

list_jobs_input

See ListJobsInput

list_tags_for_resource_input

See ListTagsForResourceInput

list_webhooks_input

See ListWebhooksInput

start_deployment_input

See StartDeploymentInput

start_job_input

See StartJobInput

stop_job_input

See StopJobInput

tag_resource_input

See TagResourceInput

untag_resource_input

See UntagResourceInput

update_app_input

See UpdateAppInput

update_branch_input

See UpdateBranchInput

update_domain_association_input

See UpdateDomainAssociationInput

update_webhook_input

See UpdateWebhookInput

Structs

CreateAppInput

The request structure used to create apps in Amplify.

CreateBackendEnvironmentInput

The request structure for the backend environment create request.

CreateBranchInput

The request structure for the create branch request.

CreateDeploymentInput

The request structure for the create a new deployment request.

CreateDomainAssociationInput

The request structure for the create domain association request.

CreateWebhookInput

The request structure for the create webhook request.

DeleteAppInput

Describes the request structure for the delete app request.

DeleteBackendEnvironmentInput

The request structure for the delete backend environment request.

DeleteBranchInput

The request structure for the delete branch request.

DeleteDomainAssociationInput

The request structure for the delete domain association request.

DeleteJobInput

The request structure for the delete job request.

DeleteWebhookInput

The request structure for the delete webhook request.

GenerateAccessLogsInput

The request structure for the generate access logs request.

GetAppInput

The request structure for the get app request.

GetArtifactUrlInput

Returns the request structure for the get artifact request.

GetBackendEnvironmentInput

The request structure for the get backend environment request.

GetBranchInput

The request structure for the get branch request.

GetDomainAssociationInput

The request structure for the get domain association request.

GetJobInput

The request structure for the get job request.

GetWebhookInput

The request structure for the get webhook request.

ListAppsInput

The request structure for the list apps request.

ListArtifactsInput

Describes the request structure for the list artifacts request.

ListBackendEnvironmentsInput

The request structure for the list backend environments request.

ListBranchesInput

The request structure for the list branches request.

ListDomainAssociationsInput

The request structure for the list domain associations request.

ListJobsInput

The request structure for the list jobs request.

ListTagsForResourceInput

The request structure to use to list tags for a resource.

ListWebhooksInput

The request structure for the list webhooks request.

StartDeploymentInput

The request structure for the start a deployment request.

StartJobInput

The request structure for the start job request.

StopJobInput

The request structure for the stop job request.

TagResourceInput

The request structure to tag a resource with a tag key and value.

UntagResourceInput

The request structure for the untag resource request.

UpdateAppInput

The request structure for the update app request.

UpdateBranchInput

The request structure for the update branch request.

UpdateDomainAssociationInput

The request structure for the update domain association request.

UpdateWebhookInput

The request structure for the update webhook request.