Module aws_sdk_amplify::model

Modules

app

See App

artifact

See Artifact

auto_branch_creation_config

See AutoBranchCreationConfig

backend_environment

See BackendEnvironment

branch

See Branch

custom_rule

See CustomRule

domain_association

See DomainAssociation

job

See Job

job_summary

See JobSummary

production_branch

See ProductionBranch

step

See Step

sub_domain

See SubDomain

sub_domain_setting

See SubDomainSetting

webhook

See Webhook

Structs

App

Represents the different branches of a repository for building, deploying, and

Artifact

Describes an artifact.

AutoBranchCreationConfig

Describes the automated branch creation configuration.

BackendEnvironment

Describes the backend environment for an Amplify app.

Branch

The branch for an Amplify app, which maps to a third-party repository branch.

CustomRule

Describes a custom rewrite or redirect rule.

DomainAssociation

Describes a domain association that associates a custom domain with an Amplify app.

Job

Describes an execution job for an Amplify app.

JobSummary

Describes the summary for an execution job for an Amplify app.

ProductionBranch

Describes the information about a production branch for an Amplify app.

Step

Describes an execution step, for an execution job, for an Amplify app.

SubDomain

The subdomain for the domain association.

SubDomainSetting

Describes the settings for the subdomain.

Webhook

Describes a webhook that connects repository events to an Amplify app.

Enums

DomainStatus

JobStatus

JobType

Platform

Stage