Module aws_sdk_amplify::operation

Structs

CreateApp

Creates a new Amplify app.

CreateBackendEnvironment

Creates a new backend environment for an Amplify app.

CreateBranch

Creates a new branch for an Amplify app.

CreateDeployment

Creates a deployment for a manually deployed Amplify app. Manually deployed apps are

CreateDomainAssociation

Creates a new domain association for an Amplify app. This action associates a custom

CreateWebhook

Creates a new webhook on an Amplify app.

DeleteApp

Deletes an existing Amplify app specified by an app ID.

DeleteBackendEnvironment

Deletes a backend environment for an Amplify app.

DeleteBranch

Deletes a branch for an Amplify app.

DeleteDomainAssociation

Deletes a domain association for an Amplify app.

DeleteJob

Deletes a job for a branch of an Amplify app.

DeleteWebhook

Deletes a webhook.

GenerateAccessLogs

Returns the website access logs for a specific time range using a presigned URL.

GetApp

Returns an existing Amplify app by appID.

GetArtifactUrl

Returns the artifact info that corresponds to an artifact id.

GetBackendEnvironment

Returns a backend environment for an Amplify app.

GetBranch

Returns a branch for an Amplify app.

GetDomainAssociation

Returns the domain information for an Amplify app.

GetJob

Returns a job for a branch of an Amplify app.

GetWebhook

Returns the webhook information that corresponds to a specified webhook ID.

ListApps

Returns a list of the existing Amplify apps.

ListArtifacts

Returns a list of artifacts for a specified app, branch, and job.

ListBackendEnvironments

Lists the backend environments for an Amplify app.

ListBranches

Lists the branches of an Amplify app.

ListDomainAssociations

Returns the domain associations for an Amplify app.

ListJobs

Lists the jobs for a branch of an Amplify app.

ListTagsForResource

Returns a list of tags for a specified Amazon Resource Name (ARN).

ListWebhooks

Returns a list of webhooks for an Amplify app.

StartDeployment

Starts a deployment for a manually deployed app. Manually deployed apps are not

StartJob

Starts a new job for a branch of an Amplify app.

StopJob

Stops a job that is in progress for a branch of an Amplify app.

TagResource

Tags the resource with a tag key and value.

UntagResource

Untags a resource with a specified Amazon Resource Name (ARN).

UpdateApp

Updates an existing Amplify app.

UpdateBranch

Updates a branch for an Amplify app.

UpdateDomainAssociation

Creates a new domain association for an Amplify app.

UpdateWebhook

Updates a webhook.