Module aws_sdk_amplify::output

Modules

create_app_output

See CreateAppOutput

create_backend_environment_output

See CreateBackendEnvironmentOutput

create_branch_output

See CreateBranchOutput

create_deployment_output

See CreateDeploymentOutput

create_domain_association_output

See CreateDomainAssociationOutput

create_webhook_output

See CreateWebhookOutput

delete_app_output

See DeleteAppOutput

delete_backend_environment_output

See DeleteBackendEnvironmentOutput

delete_branch_output

See DeleteBranchOutput

delete_domain_association_output

See DeleteDomainAssociationOutput

delete_job_output

See DeleteJobOutput

delete_webhook_output

See DeleteWebhookOutput

generate_access_logs_output

See GenerateAccessLogsOutput

get_app_output

See GetAppOutput

get_artifact_url_output

See GetArtifactUrlOutput

get_backend_environment_output

See GetBackendEnvironmentOutput

get_branch_output

See GetBranchOutput

get_domain_association_output

See GetDomainAssociationOutput

get_job_output

See GetJobOutput

get_webhook_output

See GetWebhookOutput

list_apps_output

See ListAppsOutput

list_artifacts_output

See ListArtifactsOutput

list_backend_environments_output

See ListBackendEnvironmentsOutput

list_branches_output

See ListBranchesOutput

list_domain_associations_output

See ListDomainAssociationsOutput

list_jobs_output

See ListJobsOutput

list_tags_for_resource_output

See ListTagsForResourceOutput

list_webhooks_output

See ListWebhooksOutput

start_deployment_output

See StartDeploymentOutput

start_job_output

See StartJobOutput

stop_job_output

See StopJobOutput

tag_resource_output

See TagResourceOutput

untag_resource_output

See UntagResourceOutput

update_app_output

See UpdateAppOutput

update_branch_output

See UpdateBranchOutput

update_domain_association_output

See UpdateDomainAssociationOutput

update_webhook_output

See UpdateWebhookOutput

Structs

CreateAppOutput

CreateBackendEnvironmentOutput

The result structure for the create backend environment request.

CreateBranchOutput

The result structure for create branch request.

CreateDeploymentOutput

The result structure for the create a new deployment request.

CreateDomainAssociationOutput

The result structure for the create domain association request.

CreateWebhookOutput

The result structure for the create webhook request.

DeleteAppOutput

The result structure for the delete app request.

DeleteBackendEnvironmentOutput

The result structure of the delete backend environment result.

DeleteBranchOutput

The result structure for the delete branch request.

DeleteDomainAssociationOutput

DeleteJobOutput

The result structure for the delete job request.

DeleteWebhookOutput

The result structure for the delete webhook request.

GenerateAccessLogsOutput

The result structure for the generate access logs request.

GetAppOutput

GetArtifactUrlOutput

Returns the result structure for the get artifact request.

GetBackendEnvironmentOutput

The result structure for the get backend environment result.

GetBranchOutput

GetDomainAssociationOutput

The result structure for the get domain association request.

GetJobOutput

GetWebhookOutput

The result structure for the get webhook request.

ListAppsOutput

The result structure for an Amplify app list request.

ListArtifactsOutput

The result structure for the list artifacts request.

ListBackendEnvironmentsOutput

The result structure for the list backend environments result.

ListBranchesOutput

The result structure for the list branches request.

ListDomainAssociationsOutput

The result structure for the list domain association request.

ListJobsOutput

The maximum number of records to list in a single response.

ListTagsForResourceOutput

The response for the list tags for resource request.

ListWebhooksOutput

The result structure for the list webhooks request.

StartDeploymentOutput

The result structure for the start a deployment request.

StartJobOutput

The result structure for the run job request.

StopJobOutput

The result structure for the stop job request.

TagResourceOutput

The response for the tag resource request.

UntagResourceOutput

The response for the untag resource request.

UpdateAppOutput

The result structure for an Amplify app update request.

UpdateBranchOutput

The result structure for the update branch request.

UpdateDomainAssociationOutput

The result structure for the update domain association request.

UpdateWebhookOutput

The result structure for the update webhook request.